requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.6.0
backports.cached-property==1.0.2
click==8.1.3
Django==4.0.5
graphql-core==3.2.1
h11==0.13.0
idna==3.3
pycodestyle==2.8.0
Pygments==2.12.0
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.20.4
strawberry-graphql==0.115.0
strawberry-graphql-django==0.3.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
